Fix cron changes and migration filename regex
What?
=====
- Fixed the changelog function that gets cron changes to call the
correct do command (get-cron-files instead of get_cron_files)

- Fixed the cron change processing to only consider lines that start
with 'apps/' because all the cron files should be under that directory
and it's possible to get a lot of weird output from warnings and things.

- Fixed the migration filename regex to allow filenames with dashes
in addition to underscores.

Why?
=====
The changelog for eventboard 19.36 was missing cron changes and a
new migration.