Add JIRA changelog and better support for subtasks
What?
=====
- Added a JIRA changelog section to the normal changelog command, which
contains the issue type and nesting for subtasks as well as the
assignee.

- Skip linking subtasks to the release ticket.

Why?
=====
JIRA often contains more info than the git PR title, like a better
description, issue type, and assignee. We often have to look up these
details manually.

We use subtasks to break up a story, but they are not deployable units
on their own and so don't need to be linked to the release ticket. The
subtasks often make the release ticket confusingly huge.